Refactor SQL connection and add debugging in barcode_generator
This commit eliminates the use of the sql_connect_lato function in the barcode_generator.py file, opting to use only the sql_connect. Additionally, it improves the connection to the SQL database by getting the engine through connect(), and adding debugging lines to print the type of the engine object and the dataframe.